containers: Move individual subdirectories from full_image
Newer docker (>=23.0) uses BuildKit as the default builder. That behaves
differently when copying subdirectories to symlinks on the target, and
will bail out with error message

"ERROR: failed to solve: cannot copy to non-directory: ..."

Breaking the copying up into individual subdirectories fixes this issue
as it eliminates ambiguity in the order of operations that mey influence
each other due to the symbolic links.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>