settings: update to k8s version 1.32
Update to k8s version 1.32 to be compatible with latest EKS.
Starting with version 1.30 EKS started to use AL2023 instead of AL2 as
the default ami for nodes. On AL2023 `aws-nitro-enclaves-cli` is not
part of the extras repository anymore, but can be installed from the
core repositories.

While here, move the dependency installation into cloud-config section.

This commit is partly based on a previous PR by dornseid@amazon.com [1]

[1] https://github.com/aws/aws-nitro-enclaves-with-k8s/pull/32

Signed-off-by: Leonard Foerster <foersleo@amazon.com>